Justin's modeling rough draft
This file was a rough draft of the criteria mentioned in the assignment description. I did not do any analysis or writing portion yet, just wanted to do the code for now before we collab with each other. I  was also going to put all of the files into one data frame to be trained better, but my laptop was out of storage. It would be an easy fix though because I made the variable meant for us to try testing the data frames together if we wanted to.